requirements.txt
numpy
opencv-python
nibabel
totalsegmentator
tqdm
torch
torchvision
scikit-learn
matplotlib
seaborn
pandas
Pillow
gradio
tensorboard